generate READMEs from POD.
use DZP:ReadmeAnyFromPod to do the hard work;
generate text README in the dist;
maintain a README.html in the source, for the benefit of Github;
fix typo & authorship

Why the confusion about the author?
I did part of the early research & implementation for this project at
$work, but most of the polishing and learning about dzil in my own
time.  It doesn't matter to me who "owns the copyright" while the
software is Free.